requirements.txt
attrs==20.2.0
boto3==1.14.14
botocore==1.17.14
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
docutils==0.15.2
Fiona==1.8.17
future==0.18.2
geopandas==0.8.1
gitdb==4.0.5
GitPython==3.1.3
googledrivedownloader==0.4
idna==2.9
imgviz==1.1.0
jmespath==0.10.0
kiwisolver==1.2.0
labelme==4.2.10
matplotlib==3.2.1
munch==2.5.0
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.1.2
Pillow==7.1.2
pyparsing==2.4.7
pyproj==2.6.1.post1
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
QtPy==1.9.0
requests==2.23.0
s3transfer==0.3.3
Shapely==1.7.1
six==1.14.0
smmap==3.0.4
termcolor==1.1.0
treelib==1.6.1
urllib3==1.25.9
wget==3.2
torch==1.2.0
torchvision==0.4.0
pycocotools==2.0.2